job-agent-mcp/requirements.txt
streamlit
sentence-transformers
scikit-learn
pandas
numpy
requests
spacy
fastapi
uvicorn
mcp
google-adk
PyPDF2
python-docx
aiohttp